+++ src/player-quest.h
@@ -30,6 +30,7 @@ bool quest_check(struct player *p, const struct monster *m);
 struct quest *quest_take(struct player *p, int type, const char *name,
 						 struct monster_race *race, int number);
 bool quest_check_arrival(struct player *p, int town);
+bool quest_check_item(struct player *p, const struct object *obj);
 struct quest *quest_carried(struct player *p, bool done);
 void quest_hand_back(struct player *p, struct quest *q);
 extern struct file_parser quests_parser;